backend/requirements.txt
# Haw Chat Backend - Python Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for AI API integrations)
httpx>=0.26.0

# Async support
anyio>=4.0.0

# Configuration management
pyyaml>=6.0

# LangChain for agent framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0

# Optional: For future AI integrations
# anthropic>=0.21.0  # Claude API (already in langchain-anthropic)
# openai>=1.12.0     # OpenAI API (already in langchain-openai)
